requirements.txt
# AlphaQCM 完整依赖包
# 包含运行整个系统所需的全部包（基于完整代码分析）

# 核心科学计算
torch==1.13.1
numpy==1.23.5
pandas==1.5.3
scipy==1.10.1

# 机器学习
scikit-learn==1.3.1
lightgbm==4.1.0

# 强化学习
gymnasium==0.29.1

# 数据处理
pyarrow
PyYAML==6.0.1

# 工具
tqdm==4.66.1
